requirements.txt
# Core ETL dependencies
pandas==2.2.2
pymysql==1.1.1
python-dotenv==1.0.1

# Google Cloud dependencies
google-cloud-storage==2.18.0
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
functions-framework==3.8.1

# Data processing (compatible version for M1/M2 Macs)
pyarrow>=12.0.0,<16.0.0
openpyxl==3.1.2

# Development dependencies
jupyter==1.1.1
ipykernel==6.29.5
black==24.8.0
pylint==3.2.6

# Monitoring and logging
google-cloud-monitoring==2.21.0
google-cloud-logging==3.11.0